requirements.txt
google-play-scraper
pandas
langdetect
bertopic
google-cloud-aiplatform
groq
scikit-learn
umap-learn
hdbscan
openai
sentence-transformers
python-dotenv
